move ME0DigiProducers to plugins
